+++ lua/core/commands.lua
@@ -12,6 +12,8 @@ M.defaults = {
   ]],
   -- :LvimInfo
   [[command! LvimInfo lua require('core.info').toggle_popup(vim.bo.filetype)]],
+  [[ command! LvimCacheReset lua require('utils').lvim_cache_reset() ]],
+  [[ command! LvimUpdate lua require('utils').lvim_update() ]],
 }
 
 M.load = function(commands)
